requirements.txt
Flask-SQLAlchemy>=2.4.1
SQLAlchemy>=1.3.15
PyYAML>=5.3.1
Flask>=1.1.1
pymysql>=0.9.3

numpy>=1.18.4
pandas>=1.0.3
boardgamegeek2>=1.0.1
argparse>=1.1
boto3>=1.13.3
boto>=2.49.0
scikit-learn>=0.23.1
pytest>=5.4.2
botocore>=1.15.39
